chemical_anomaly_detection/requirements.txt
crewai
crewai-tools
qdrant-client
opencv-python
librosa
sentence-transformers
numpy
python-dotenv
pydantic
soundfile
transformers
accelerate
torchaudio
torch
torchvision
panns_inference
pandas
moviepy
litellm
fastapi
uvicorn[standard]
python-multipart